Fix bug with Fixpoint API key on requests (#45)
I changed the `Requester.api_key` from a property to a method, but
missed updating where we referenced the api key when constructing
requests. I was trying to set the auth header to the string value of the
function, but it should be the string value of the result of the
function.